+++ app/src/main/java/com/ssomcompany/ssomclient/common/Util.java
@@ -159,10 +159,10 @@ public class Util {
      * @param dp A value in dp (density independent pixels) unit. Which we need to convert into pixels
      * @return A float value to represent px equivalent to dp depending on device density
      */
-    public static float convertDpToPixel(float dp){
+    public static int convertDpToPixel(float dp){
         Resources resources = BaseApplication.getInstance().getResources();
         DisplayMetrics metrics = resources.getDisplayMetrics();
-        return dp * (metrics.densityDpi / DisplayMetrics.DENSITY_DEFAULT);
+        return (int) dp * (metrics.densityDpi / DisplayMetrics.DENSITY_DEFAULT);
     }
 
     /**
